refactor util classes
