+++ platforms/fabric/src/main/java/com/dfsek/terra/fabric/TerraFabricPlugin.java
@@ -41,6 +41,7 @@ import com.dfsek.terra.fabric.generation.TerraBiomeSource;
 import com.dfsek.terra.fabric.handle.FabricItemHandle;
 import com.dfsek.terra.fabric.handle.FabricWorldHandle;
 import com.dfsek.terra.fabric.mixin.access.GeneratorTypeAccessor;
+import com.dfsek.terra.fabric.util.FabricUtil;
 import com.dfsek.terra.profiler.Profiler;
 import com.dfsek.terra.profiler.ProfilerImpl;
 import com.dfsek.terra.registry.exception.DuplicateEntryException;
@@ -308,7 +309,7 @@ public class TerraFabricPlugin implements TerraPlugin, ModInitializer {
     @Addon("Terra-Fabric")
     @Author("Terra")
     @Version("1.0.0")
-    static final class FabricAddon extends TerraAddon implements EventListener {
+    public static final class FabricAddon extends TerraAddon implements EventListener {
 
         private final TerraPlugin main;
